ecommerce/requirement.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
cssselect2==0.7.0
distlib==0.3.7
Django==4.2.4
django-appconf==1.0.5
django-autoslug==1.9.9
django-shopping-cart==0.1
django-tinymce==3.6.1
django-widget-tweaks==1.5.0
fibrenest-db-models==21.11
filelock==3.13.1
greenlet==3.0.1
html5lib==1.1
idna==3.4
lxml==4.9.3
oscrypto==1.3.0
Pillow==10.1.0
platformdirs==3.11.0
pycparser==2.21
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
pypdf==3.17.0
pypng==0.20220715.0
python-bidi==0.4.2
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
such-distribution==2.0
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uritools==4.0.2
urllib3==2.0.7
virtualenv==20.24.6
webencodings==0.5.1
xhtml2pdf2==0.2.12
